% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanova.R
\name{stanova}
\alias{stanova}
\title{Estimate ANOVA-type models with rstanarm}
\usage{
stanova(formula, data, model_fun, ..., check_contrasts = "contr.bayes")
}
\arguments{
\item{formula}{a formula describing the statistical model to be estimated.
Passed to the \code{rstanarm::stan_} function specified in \code{model_fun}.}

\item{data}{\code{data.frame} containing the data.}

\item{model_fun}{character string identifying the \code{rstanarm} function that
should be used for fitting (omitting the \code{stan_} prefix) such as \code{"lm"} or
\code{"lmer"}.}

\item{...}{further arguments passed to the specified \code{rstanarm::stan_}
function.}

\item{check_contrasts}{\code{character} string (of length 1) denoting a contrast
function which should be assigned to all \code{character} and \code{factor} variables
in the model (as long as the specified contrast is not the global default).
Default is \link{contr.bayes}.}
}
\description{
Estimate ANOVA-type models with rstanarm
}
\examples{

library("rstanarm")
m1 <- stanova(yield ~ block + N*P*K, npk, prior = R2(0.5))
m1
summary(m1)

fit_warp <- stanova(breaks ~ wool * tension, data = warpbreaks, prior = R2(0.5))
summary(fit_warp)

####

data("Machines", package = "MEMSS")

m_machines <- stanova(score ~ Machine + (Machine|Worker),
                      model_fun = "lmer",
                      data=Machines, chains = 2, iter = 500)
summary(m_machines)

## negative binomial
fit6a <- stanova(Days ~ Sex/(Age + Eth*Lrn), data = MASS::quine,
                 model_fun = "glm.nb",
                 link = "log", prior_aux = exponential(1.5),
                 chains = 2, iter = 200) # for speed of example only
summary(fit6a)

## beta regression (fails somehow)
data("GasolineYield", package = "betareg")
str(GasolineYield)

fit_beta <- stanova(yield ~ batch + temp, data = GasolineYield,
                    model_fun = "betareg",
                    link = "logit", link.phi = "log",
                    chains = 2, iter = 500)
summary(fit_beta)
}
