% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanova_brm.R
\name{stanova_brm}
\alias{stanova_brm}
\title{Estimate ANOVA-type models with brms}
\usage{
stanova_brm(formula, data, model_fun, ..., check_contrasts = "contr.bayes")
}
\arguments{
\item{formula}{a formula describing the statistical model to be estimated.
Passed to the \code{rstanarm::stan_} function specified in \code{model_fun}.}

\item{data}{\code{data.frame} containing the data.}

\item{model_fun}{character string identifying the \code{rstanarm} function that
should be used for fitting (omitting the \code{stan_} prefix) such as \code{"glm"} or
\code{"glmer"}.}

\item{...}{further arguments passed to the \code{rstanarm} function used for
fitting. Typical arguments are \code{prior}, \code{prior_intercept}, \code{chain}, \code{iter},
or \code{core}.}

\item{check_contrasts}{\code{character} string (of length 1) denoting a contrast
function or a contrast function which should be assigned to all \code{character}
and \code{factor} variables in the model (as long as the specified contrast is
not the global default). Default is \link{contr.bayes}. Set to \code{NULL} to disable
the check.}
}
\description{
Estimate ANOVA-type models with brms
}
\examples{

\dontrun{
fit_warp_brms <- stanova_brm(breaks ~ wool * tension, data = warpbreaks,
                             chains = 2, iter = 1000)
summary(fit_warp_brms)
}

\dontrun{
### binomial GLMM

data("stroop", package = "afex")
## for frequentist analysis see:
### https://cran.r-project.org/package=afex/vignettes/afex_analysing_accuracy_data.html
library("tidyverse")
stroop_e1 <- stroop \%>\%
  filter(!is.na(acc)) \%>\%
  filter(study == "1") \%>\%
  droplevels() \%>\%
  group_by(condition, congruency, pno) \%>\%
  summarise(
    succ = sum(acc),
    acc = mean(acc),
    n = n())

bglmm_stroop <- stanova_brm(
  succ | trials(n) ~ congruency*condition + (congruency*condition|pno),
  data = stroop_e1, family = binomial,
  chains = 2, iter = 1000
)
summary(bglmm_stroop)
}
}
