% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanova_samples.R
\name{stanova_samples}
\alias{stanova_samples}
\title{Extract stanova difference from intercept samples}
\usage{
stanova_samples(
  object,
  terms,
  return = c("array", "matrix", "data.frame"),
  dimension_chain = 3L,
  ...
)
}
\arguments{
\item{object}{Fitted model. Currently only tested for \code{stanova} objects but
likely supports \code{rstanarm} and potentially other objects as well.}

\item{terms}{optional character vector denoting a subset of model terms for
which difference from intercept be returned.}

\item{return}{\code{character} string denoting format in which samples should be
returned. Allowed values are \code{c("array", "matrix", "data.frame")}, possibly
abbreviated.}

\item{dimension_chain}{Scalar integer. If \code{return = "array"}, determines the
dimension of the chain. The default \code{3} means chain is the third dimension
of the returned array. Value should be between 1 and 3.}
}
\description{
Extract stanova difference from intercept samples
}
