% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanova.R
\name{stanova}
\alias{stanova}
\title{Estimate ANOVA-type models with rstanarm}
\usage{
stanova(formula, data, model_fun, ..., check_contrasts = "contr.bayes")
}
\arguments{
\item{formula}{a formula describing the statistical model to be estimated.
Passed to the \code{rstanarm::stan_} function specified in \code{model_fun}.}

\item{data}{\code{data.frame} containing the data.}

\item{model_fun}{character string identifying the \code{rstanarm} function that
should be used for fitting (omitting the \code{stan_} prefix) such as \code{"glm"} or
\code{"lmer"}.}

\item{...}{further arguments passed to the \code{rstanarm} function used for
fitting. Typical arguments are \code{prior}, \code{prior_intercept}, \code{chain}, \code{iter},
or \code{core}.}

\item{check_contrasts}{\code{character} string (of length 1) denoting a contrast
function which should be assigned to all \code{character} and \code{factor} variables
in the model (as long as the specified contrast is not the global default).
Default is \link{contr.bayes}. Set to \code{NULL} to disable the check.}
}
\description{
Estimate ANOVA-type models with rstanarm
}
\examples{

fit_warp <- stanova(breaks ~ wool * tension, data = warpbreaks,
                    prior = rstanarm::student_t(3, 0, 3, autoscale = FALSE),
                    model_fun = "glm",
                    chains = 2, iter = 500)
summary(fit_warp)

####

data("Machines", package = "MEMSS")

## note: model_fun = "lmer" only works after library("rstanarm")
## but you can simply use model_fun = "glmer" with family = "gaussian"
m_machines <- stanova(score ~ Machine + (Machine|Worker),
                      model_fun = "glmer", family = "gaussian",
                      data=Machines, chains = 2, iter = 500)
summary(m_machines)

\dontrun{
## negative binomial
## requires attaching rstanarm
library("rstanarm")
fit6a <- stanova(Days ~ Sex/(Age + Eth*Lrn), data = MASS::quine,
                 model_fun = "glm.nb",
                 link = "log", prior_aux = exponential(1.5),
                 chains = 2, iter = 200) # for speed of example only
summary(fit6a)
}

}
