% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanova_lmer.R
\name{stanova_lmer}
\alias{stanova_lmer}
\title{Estimate ANOVA-type models with rstanarm}
\usage{
stanova_lmer(formula, data, check_contrasts = "contr.bayes", ...)
}
\arguments{
\item{formula}{a formula describing the full mixed-model to be fitted. Passed
to \code{\link[rstanarm:stan_lmer]{rstanarm::stan_lmer()}}.}

\item{data}{\code{data.frame} containing the data.}

\item{check_contrasts}{\code{character} string (of length 1) denoting a contrast
function which should be assigned to all \code{character} and \code{factor} variables
in the model (as long as the specified contrast is not the global default).
Default is \link{contr.bayes}.}

\item{...}{further arguments passed to \code{\link[rstanarm:stan_lmer]{rstanarm::stan_lmer}}.}
}
\description{
Estimate ANOVA-type models with rstanarm
}
\examples{

data("Machines", package = "MEMSS")

m_machines <- stanova_lmer(score ~ Machine + (Machine|Worker),
                           data=Machines, chains = 2, iter = 500)

out_list <- summary(m_machines)
str(out_list)

out_list

}
