% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stanova_samples.R
\name{stanova_samples}
\alias{stanova_samples}
\alias{stanova_samples.stanova}
\title{Posterior samples representing difference from intercept}
\usage{
stanova_samples(object, ...)

\method{stanova_samples}{stanova}(
  object,
  diff_intercept = TRUE,
  terms,
  return = c("array", "matrix", "data.frame"),
  dimension_chain = 3L,
  ...
)
}
\arguments{
\item{object}{Fitted model. Currently only support for \code{stanova} objects but
likely supports \code{rstanarm} and potentially other objects as well.}

\item{...}{currently ignored.}

\item{diff_intercept}{logical. If TRUE (the default) samples for factor
levels represent the difference from the intercept. If FALSE, marginal
means of factor levels are returned.}

\item{terms}{optional character vector denoting a subset of model terms for
which difference from intercept be returned.}

\item{return}{\code{character} string denoting format in which samples should be
returned. Allowed values are \code{c("array", "matrix", "data.frame")}, possibly
abbreviated.}

\item{dimension_chain}{Scalar integer. If \code{return = "array"}, determines the
dimension of the chain. The default \code{3} means chain is the third dimension
of the returned array. Value should be between 1 and 3.}
}
\description{
Posterior samples representing difference from intercept
}
